refactor(auth): keep resource access within complexity limits
