+++ packages/plugin-sdk/src/standard-content-plugin.ts
@@ -1,6 +1,5 @@
 import { definePluginAdminResources, type AdminResourceDefinition } from './admin-resources.js';
 import { definePluginContentTypes, type ContentTypeDefinition } from './content-types.js';
-import type { SessionAccessSnapshot } from './session-access.js';
 import {
   definePluginActions,
   definePluginAuditEvents,
@@ -26,38 +25,11 @@ export type StandardContentPluginActionIds<TPluginId extends string = string> = 
 export type StandardContentPluginActionOptions = Readonly<{
   legacyAliases?: Partial<Readonly<Record<StandardContentPluginActionName, readonly string[]>>>;
 }>;
-export type StandardContentAccessCapabilities = Readonly<{
-  isResolved: boolean;
-  canRead: boolean;
-  canCreate: boolean;
-  canUpdate: boolean;
-  canDelete: boolean;
-}>;
-
-export type StandardContentResourceAccess = Readonly<Record<string, boolean>>;
-
-export const resolveStandardContentAccessCapabilities = (
-  pluginId: string,
-  snapshot: SessionAccessSnapshot,
-  resourceAccess: StandardContentResourceAccess = {}
-): StandardContentAccessCapabilities => {
-  const hasAssignedModule = snapshot.assignedModules.includes(pluginId);
-  const actions = new Set(snapshot.permissionActions);
-  const unscopedActions = new Set(snapshot.unscopedPermissionActions ?? []);
-  const allows = (action: string): boolean =>
-    snapshot.isResolved && hasAssignedModule && actions.has(`${pluginId}.${action}`);
-  const allowsMutation = (action: 'update' | 'delete'): boolean => {
-    const actionId = `${pluginId}.${action}`;
-    return allows(action) && (unscopedActions.has(actionId) || resourceAccess[actionId] === true);
-  };
-  return {
-    isResolved: snapshot.isResolved,
-    canRead: allows('read'),
-    canCreate: allows('create'),
-    canUpdate: allowsMutation('update'),
-    canDelete: allowsMutation('delete'),
-  };
-};
+export { resolveStandardContentAccessCapabilities } from './standard-content-access.js';
+export type {
+  StandardContentAccessCapabilities,
+  StandardContentResourceAccess,
+} from './standard-content-access.js';
 
 export type StandardContentAdminResourceOptions = Readonly<{
   pluginId: string;
