+++ packages/sva-mainserver/src/server/mutation-principal-authorization.ts
@@ -159,7 +159,6 @@ type AuthorizationAggregate = {
   shadowDifference: boolean;
 };
 type ProviderDecision = Awaited<ReturnType<typeof authorizeMainserverDataProviderAccess>>;
-export type MainserverResourceAccess = Readonly<Record<string, boolean>>;
 const mergeProviderDecision = (
   aggregate: AuthorizationAggregate,
   decision: ProviderDecision
@@ -317,41 +316,4 @@ export const authorizeMainserverExistingContent = async (input: {
       : {}),
     ...(authorization.shadowDifference ? { shadowDifference: true } : {}),
   });
-};
-
-export const resolveMainserverResourceAccess = async (input: {
-  readonly actor: MainserverMutationActor;
-  readonly actions: readonly string[];
-  readonly contentType: string;
-  readonly item: DataProviderBearingItem | undefined;
-}): Promise<MainserverResourceAccess> => {
-  const denied = Object.fromEntries(input.actions.map((action) => [action, false]));
-  const permissions = await loadMutationPermissions(input.actor);
-  if (!permissions.ok) return denied;
-
-  const dataProviderId = input.item?.dataProvider?.id?.trim() ?? '';
-  const decisions = await Promise.all(
-    input.actions.map(async (action) => {
-      const decision = await authorizeMainserverDataProviderAccess({
-        instanceId: input.actor.instanceId,
-        keycloakSubject: input.actor.keycloakSubject,
-        actorAccountId: input.actor.actorAccountId,
-        actingPrincipalType: input.actor.mutationPrincipalContext.actingPrincipalType,
-        credentialFingerprint: input.actor.mutationPrincipalContext.credentialFingerprint,
-        ...(input.actor.mutationPrincipalContext.contentAuthorPolicy
-          ? { contentAuthorPolicy: input.actor.mutationPrincipalContext.contentAuthorPolicy }
-          : {}),
-        ...(input.actor.activeOrganizationId
-          ? { activeOrganizationId: input.actor.activeOrganizationId }
-          : {}),
-        action,
-        contentType: input.contentType,
-        contentId: input.item?.id,
-        permissions: permissions.permissions,
-        dataProviderId,
-      });
-      return [action, decision.allowed] as const;
-    })
-  );
-  return Object.fromEntries(decisions);
 };
